+++ src/app/shell/selectors.ts
@@ -4,3 +4,4 @@ export const isPhonePortraitSelector = (state: RootState) => state.shell.isPhone
 export const querySelector = (state: RootState) => state.shell.searchQuery;
 export const searchQueryVersionSelector = (state: RootState) => state.shell.searchQueryVersion;
 export const bungieAlertsSelector = (state: RootState) => state.shell.bungieAlerts;
+export const searchResultsOpenSelector = (state: RootState) => state.shell.searchResultsOpen;